WO:Create a special datetime widget which takes care about date AND time
Finally it wasn't possible to use the same date model for date AND time due to dateRangeSlider synchronization... so we introduced two additional properties on model : startTime and stopTime (as it was done long time before). Not the best solution of my life but had some problems with winter/summer times, so dateRangeSlider jumped to day before/after ... maybe pass to moment lib one day. Don't really want to touch this for now..